move ticker pkg
